// src/pageObjects/objects/athena/motor/ph-driving-licence-details/aliases.generated.ts
export const pageMeta = {
  pageKey: "athena.motor.ph-driving-licence-details",
  urlPath: "/journey/show/product/AnnualMotorInsurance/process/nb/drivingLicenceScan/69b442a3e1389c2e75667e1d",
  urlRe: /\/journey\/show\/product\/AnnualMotorInsurance\/process\/nb\/drivingLicenceScan\/[a-z0-9]+/i,
  title: "Driving licence details",
  titleRe: new RegExp("Driving licence details", "i"),
} as const;